kafka: move the common parts of the destination driver to kafka-internal.c/h
