+++ modules/kafka/kafka-dest-worker.h
@@ -27,14 +27,7 @@
 
 #include "logthrdest/logthrdestdrv.h"
 
-typedef struct _KafkaDestWorker
-{
-  LogThreadedDestWorker super;
-  struct iv_timer poll_timer;
-  GString *key;
-  GString *message;
-  GString *topic_name_buffer;
-} KafkaDestWorker;
+typedef struct _KafkaDestWorker KafkaDestWorker;
 
 LogThreadedDestWorker *kafka_dest_worker_new(LogThreadedDestDriver *owner, gint worker_index);
