+++ modules/kafka/kafka-internal.h
@@ -2,6 +2,7 @@
  * Copyright (c) 2020 Balabit
  * Copyright (c) 2020 Balazs Scheidler
  * Copyright (c) 2020 Vivin Peris
+ * Copyright (c) 2025 Hofi <hofione@gmail.com>
  *
  * This program is free software: you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by
@@ -22,17 +23,96 @@
  *
  */
 
+/*
+ * Original Kafka source implementation sourced from https://github.com/syslog-ng/syslog-ng/pull/5564.
+ * The license has been upgraded from GPL-2.0-or-later to GPL-3.0-or-later.
+ */
+
 #ifndef KAFKA_INTERNAL_H_INCLUDED
 #define KAFKA_INTERNAL_H_INCLUDED
 
 #include "logthrdest/logthrdestdrv.h"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wignored-qualifiers"
 #include <librdkafka/rdkafka.h>
+#pragma GCC diagnostic pop
+#include "kafka-dest-driver.h"
 #include "kafka-dest-worker.h"
 
-gchar *kafka_dest_worker_resolve_template_topic_name(KafkaDestWorker *self, LogMessage *msg);
+#define KAFKA_DEEP_TRACE 0
+#if SYSLOG_NG_ENABLE_DEBUG
+# define kafka_msg_debug msg_verbose
+# define kafka_msg_trace msg_verbose
+# if KAFKA_DEEP_TRACE
+#  define kafka_msg_deep_trace msg_verbose
+# else
+#  define kafka_msg_deep_trace(msg, ...)
+# endif
+#else
+# define kafka_msg_debug msg_debug
+# define kafka_msg_trace msg_trace
+# define kafka_msg_deep_trace(msg, ...)
+#endif
+
+#define TOPIC_NAME_ERROR topic_name_error_quark()
+
+typedef enum _KafkaTopicError
+{
+  TOPIC_LENGTH_ZERO,
+  TOPIC_DOT_TWO_DOTS,
+  TOPIC_EXCEEDS_MAX_LENGTH,
+  TOPIC_INVALID_PATTERN,
+} KafkaTopicError;
+
+GQuark topic_name_error_quark(void);
+
+gboolean kafka_validate_topic_name(const gchar *name, GError **error);
+gboolean kafka_conf_get_prop(const rd_kafka_conf_t *conf, const gchar *name, gchar *dest, size_t *dest_size);
+gboolean kafka_conf_set_prop(rd_kafka_conf_t *conf, const gchar *name, const gchar *value);
+gboolean kafka_apply_config_props(rd_kafka_conf_t *conf, GList *props, gchar **protected_properties,
+                                  gsize protected_properties_num);
+void kafka_log_callback(const rd_kafka_t *rkt, int level, const char *fac, const char *msg);
+
+
+/* Kafka Destination */
+
+struct _KafkaDestWorker
+{
+  LogThreadedDestWorker super;
+  struct iv_timer poll_timer;
+  GString *key;
+  GString *message;
+  GString *topic_name_buffer;
+};
+
+struct _KafkaDestDriver
+{
+  LogThreadedDestDriver super;
+
+  LogTemplateOptions template_options;
+  LogTemplate *key;
+  LogTemplate *message;
+  LogTemplate *topic_name;
+  GHashTable *topics;
+  GMutex topics_lock;
+
+  gboolean transaction_commit;
+  GList *config;
+  gchar *bootstrap_servers;
+  gchar *fallback_topic_name;
+  rd_kafka_topic_t *topic;
+  rd_kafka_t *kafka;
+  gint flush_timeout_on_shutdown;
+  gint flush_timeout_on_reload;
+  gint poll_timeout;
+  gboolean transaction_inited;
+};
+
+const gchar *kafka_dest_worker_resolve_template_topic_name(KafkaDestWorker *self, LogMessage *msg);
 rd_kafka_topic_t *kafka_dest_worker_calculate_topic_from_template(KafkaDestWorker *self, LogMessage *msg);
 rd_kafka_topic_t *kafka_dest_worker_get_literal_topic(KafkaDestWorker *self);
 rd_kafka_topic_t *kafka_dest_worker_calculate_topic(KafkaDestWorker *self, LogMessage *msg);
+rd_kafka_topic_t *kafka_dd_query_insert_topic(KafkaDestDriver *self, const gchar *name);
 gboolean kafka_dd_init(LogPipe *s);
 
 #endif
